patch 0.1.42: Dump selected bits to VCD only.
It is possible to dump a selected bit of specific register only. You may
call `avrsim` with `--dump-regs=SREG0,PORTB2,PORTB5`, for instance, to add
SREG0, PORTB2 and PORTB5 bits to the dump only. The whole register and
its bits can be combined in one call like `--dump-regs=DDRB,DDRB2,DDRB0`.
This change should help to debug protocols built on top of two-wires
interface of AVR MCU, for instance.

Signed-off-by: Dmitry Salychev <darkness.bsd@gmail.com>