patch 0.1.45: Multi-cycle instructions simulation.
Decoder of the AVRSim has been updated to skip required number of cycles
in order to simulate multi-cycle instructions correctly (excluding DES,
LD, ST and SPM instructions at the moment). There are two flags XMEGA
and REDUCED_CORE added to the MCU initialization function to help
decoder skipping a specific number of cycles.

Also a limit for maximum number of clock cycles to store in VCD file
added. There is a minor change in the LICENSE to mention MCUSim and
its parts. CMakeLists of the MCUSim has been updated to let user
build a handbook from every directory.

Signed-off-by: Dmitry Salychev <darkness.bsd@gmail.com>