patch 0.1.52: Add MCUSim board generating script draft (#15).
This new draft of the script is a work-in-progress on
ticket #15 "Ability to create simulated boards for tests".
CMake configuration file was modified to install
this "mcusim-avr-board "script together with a simulator
renamed to "mcusim-avr".

I also decided to get rid of "WARN:", "ERRO:" and other
prefixes in the output lines of the simulator.
You'll be able to find only warning messages with tiny
prefix "[!]: " and error messages with "[e]: ".
Information messages don't contain any prefixes anymore.

Signed-off-by: Dmitry Salychev <darkness.bsd@gmail.com>