patch 0.1.64: Update PWM output when buffered OCR2->TOP.
This is an interesting case when buffered value of the OCR2 Compare
Register became TOP(0xFF). There is a weird glitch on a trace without
this adjustment applied. However, PWM output generated in Phase Correct
mode of the Timer/Counter2 is supposed to be symmetrical. That's why I
think OC2 output updated this way. I hope I'll have time to build a test
PCB and measure these values on a real MCU.

Signed-off-by: Dmitry Salychev <darkness.bsd@gmail.com>