patch 0.1.0108: USART and WDT drafts (m8a).
ticket: https://trac.mcusim.org/ticket/10

There is a lot of changes regarding further USART implementation in
ATmega8A combined with Watchdog Timer draft. It is possible to reset a
simulated MCU on Watchdog Reset now (one of the reset reasons described
in https://trac.mcusim.org/ticket/27).

There is no dedicated thread to read USART data from pseudo-terminal
master at the moment, and MCU is blocked until any data arrived on the
USART Rx channel. This is not how real hardware is supposed to work and
will be changed.

USART implementation cannot also be finished until a solution for
https://trac.mcusim.org/ticket/28.

Signed-off-by: Dmitry Salychev <darkness.bsd@gmail.com>