patch 0.1.0124: Reset SPMEN bit correctly (m8a).
ticket: https://trac.mcusim.org/ticket/7

SPMEN bit of the SPMCR register enabled the SPM instruction for the next
four clock cycles and is supposed to be cleared upon completion of the
instruction, or if no SPM instruction is executed within four clock cycles.

This patch provides a possibility for the AVR MCUs to define a "RESET_SPM_F"
function (see mcusim/avr/sim/sim.h for details) and gives an ability to
ATmega8A to reset SPMEN bit within required 4 cycles.

This patch is another step on a way to support bootloaders within the
simulated AVR microcontrollers.

Signed-off-by: Dmitry Salychev <darkness.bsd@gmail.com>