patch 0.1.0126: Fix SPM page writing.
ticket: https://trac.mcusim.org/ticket/7

This patch fixes a program memory page writing from a page buffer back to the
program memory. The main problem was an incorrect usage of the buffer
index to read a page from. Page buffer is also supposed to be as huge as
the whole program memory to be addressed by the (RAMPZ:Z) pair of registers
correctly.

Signed-off-by: Dmitry Salychev <darkness.bsd@gmail.com>