patch 0.1.0138: Fix build on GNU/Linux.
ticket: https://trac.mcusim.org/ticket/38

There is an _XOPEN_SOURCE macro which is required to be defined to
compile mcusim with 'sigaction' on GNU/Linux with -std=iso9899:1999.

This is also a change to load or dump an AVR flash memory above 64 KiB.
It does not conform to Intex HEX 8-bit, i.e. memory above 64 KiB can be
loaded with '00' and '01' records, but allows mcusim to make dumps of
the flash memory and load them back correctly.

The correct way is to support types '02' and '03' of the Intel HEX 16-bit
records in future.

Signed-off-by: Dmitry Salychev <darkness.bsd@gmail.com>