patch 0.1.0140: Add thread-safe queue.
ticket: N/A

Thread-safe queue with the blocking functions to enqueue and dequeue
elements is a first step on a way to dump VCD frames in a separate thread.

This improvement is generally necessary because a function to dump VCD
frames MSIM_AVR_VCDDumpFrame() occupies up to 65-70% of the whole time
spent in MSIM_AVR_Simulate().

This result is according to the flame graphs obtained from a real simulation of
ATmega8A with a Lua module loaded and several MCU registers to be watched and
dumped into a VCD file.