patch 0.1.0146: Cache "module_tick" value.
ticket: N/A

This patch implements a simple idea of caching a value of the "module_tick"
function of a Lua model to call it later in MSIM_AVR_LUATickModels() function
periodically.

Since calling to lua_getglobal() in a loop proved to be time consuming
(this function does not benefit from the LuaJIT compiler), it is worth
replacing it by caching a resulted value.