patch 0.1.161: Improved MSIM_AVR_Init().
ticket: N/A

Function to initialize an AVR MCU has been improved significantly. It
includes all of the code which was necessary to setup an AVR MCU in
general.

It helped to remove a lot of the initialization code from the msim_main.c
and a ATmega8A XSPICE model (cfunc.mod.c).