patch 0.1.163: Add example for NGSpice.
ticket: N/A

There is an example to simulate ATmega8A within a NGSpice circuit added.
It is a proof-of-concept to obtain a transient response of a circuit
with microcontroller and its firmware loaded.

Configuration file of the MCUSim can also be used to tweak a simulated
MCU itself and obtain a VCD file with AVR I/O registers dumped.