patch 0.1.166: New PWM-to-sine example.
ticket: N/A

There is a new example to show how a PWM signal generated by ATmega8A can be
converted to a sine wave added to scripts/examples directory.

It replaces a previous example to toggle 12V switch (built on MOSFETs)
in order to remove 3rd party SPICE models for a diode and transistors from the
repository.