patch 0.1.170: Add toggle pin test for m328p.
ticket: N/A

See tests/ATMEGA328-toggle-pin for details.

This test shows why it is not a good idea to use _delay_ms() function to
perform a delay. Delay is supposed to be ~200ms, but it is ~159ms
actually.

Firmware of the test should actually be checked in a real circuit.