patch 0.1.173: Single AVR function type.
ticket: N/A

Several different MCU-specific function types were replaced with the only one
MSIM_AVRFunc which states that the function should accept a pointer to
the simulated MCU and a pointer to a structure which holds configuration
(struct MSIM_AVRConf).

It allows to add new configuration parameters to be passed to the
MCU-specific functions without re-defining the original function type
and modifying the existing implementations of the functions.