Additional fields added to MSIM_AVR_WDT structure.
ticket: N/A

Watchdog timer clocked from the separate independent on-chip oscillator in the
AVR microcontrollers with its own frequency.

There is a new "oscf" field to denote this frequency in the MSIM_AVR_WDT
structure and two utility fields "oscp" and "scnt" to denote a current
pseudo-prescaler from the simulation loop and a system clock counter to counter
simulation loop cycles accordingly.

There could be a specific bit in one of the fuse bytes to keep WDT
active all the time also. It is called "wdton" and a new field with the same
name is supposed to track it.