Unit tests changed to "firmware" tests.
It means that these "firmware" tests are based on cmocka, but they're
intended to run an AVR simulation as usual with an ability to stop and
the selected PC values and compare an existing datamem bytes with the
ones loaded from the dump files.

These dumps (in Intel HEX format) are created with help of the Atmel Studio's
simulator (and MemoryLogger plugin), which can be treated as a reference
implementation.

See for details:https://www.avrfreaks.net/comment/2715596#comment-2715596

You can use "cmake -DWITH_FIRMWARE_TESTS=True .." to generate a makefile
and run "make && make check" then. Despite the fact that there is only
one firmware test (XlingFirmware.ft.c), it's already revealed an issue
with the general purpose and I/O registers state (incorrect value of the
0x29 PIND, expected 0x08, but read 0x00).