Switched to the progmem addressed in 16-bit words.
ticket: N/A

AVR simulator of the Atmel Studio 7 addresses program memory in 16-bit
words and implemented as dynamically linked libraries generated directly
from VHDL models of the chips.

It means that this is an expected behavior and adjusting MCUSim to
follow it make it more accurate simulator in general. It'd also help
to compare the whole datamem dump (obtained from Atmel Studio simulator) with
the actual data memory in MCUSim.

Switching to the 16-bit PM has just started and checks and tests are
still failing. This is expected.

You can build MCUSim this way to see what's failing:

    $ git clone https://github.com/mcusim/MCUSim.git
    $ cd MCUSim
    $ mkdir build && cd build
    $ cmake -DWITH_CHECKS=True ..
    $ make
    $ make checks
    $ make tests