Switched to GNU General Public License 3+.
ticket: N/A

The whole project is going to be distributed under the terms of the
GNU General Public License version 3 of the License, or (at your option) any
later version.

This should make no obstacles because the code model library, mcusim.cm, and
libmsim.so can easily be used with the projects under compatible open source
licenses which are GNU GPL 3+ in case of KiCad and New BSD License in case of
NGSpice.