Update bits of the PINx registers.
ticket: N/A

Separate bits of a PINx register should be updated accoding to the
PORTx value if they were configured to output in DDRx.

It can be done by calling 'MSIM_AVR_IOUpdatePinx' function.

There is a new '-DWITH_ASAN=True' option which can be used to enable
address sanitizer with Clang 3.1+ or GCC 4.8+ only.